Added Poisonous Bat Enemy
Fixed colliders and reduce difficulties of current enemies.

Added a new bat enemy that uses striking attacks and has a chance of poisoning the player if it gets close.
The bat kills itself when it uses poison.

I've added debug statements to say when the player has been poisoned, but need to implement the health decreasing gradually.